redistributed config / Iofog logic a bit
